requirements.txt
asgiref==3.7.2
certifi==2023.5.7
charset-normalizer==3.1.0
Django==4.2.2
django-address-app==0.1.13
django-phonenumber-field==7.1.0
django-soft-delete==0.9.21
environs==9.5.0
idna==3.4
marshmallow==3.19.0
packaging==23.1
phonenumberslite==8.13.14
psycopg2==2.9.6
psycopg2-binary==2.9.6
python-dotenv==1.0.0
requests==2.31.0
sqlparse==0.4.4
structlog==23.1.0
swapper==1.3.0
typing_extensions==4.6.3
tzdata==2023.3
urllib3==2.0.3
